# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/jobs": {
            "get": {
                "description": "Returns paginated list of jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get all jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new job entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job",
                "parameters": [
                    {
                        "description": "Job object",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_job.Job"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/active": {
            "get": {
                "description": "Returns paginated list of active jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get active jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/batch": {
            "post": {
                "description": "Creates multiple jobs at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Batch create jobs",
                "parameters": [
                    {
                        "description": "List of jobs",
                        "name": "jobs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_job.Job"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes multiple jobs by IDs or JobIDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Batch delete jobs",
                "parameters": [
                    {
                        "description": "Batch delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.BatchDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/external/{job_id}": {
            "get": {
                "description": "Returns a job by its external JobID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job by external JobID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/search": {
            "get": {
                "description": "Searches jobs with filters like query, work mode, skills, salary, etc.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Search jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Currency code",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated work modes",
                        "name": "work_mode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated work types",
                        "name": "work_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated skills",
                        "name": "skills",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated sources",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum salary",
                        "name": "min_salary",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum salary",
                        "name": "max_salary",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter active jobs",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/stats": {
            "get": {
                "description": "Returns statistics about jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/jobs/{id}/deactivate": {
            "patch": {
                "description": "Marks a job as inactive",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Deactivate a job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dtos.BatchDeleteRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "job_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_job.Job": {
            "description": "Job posting information",
            "type": "object",
            "properties": {
                "application_url": {
                    "type": "string"
                },
                "bond_period": {
                    "type": "string"
                },
                "bonuses": {
                    "type": "string"
                },
                "compensation_max": {
                    "type": "integer"
                },
                "compensation_min": {
                    "type": "integer"
                },
                "contract_duration": {
                    "type": "string"
                },
                "created_at": {
                    "description": "CompanyID uint    `json:\"company_id\"`\nCompany   Company `gorm:\"foreignKey:CompanyID;constraint:OnDelete:CASCADE\" json:\"company\"`",
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "deleted_at": {
                    "description": "@swagger:ignore",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "education_required": {
                    "type": "string"
                },
                "experience_max": {
                    "type": "integer"
                },
                "experience_min": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "interview_mode": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "job_id": {
                    "type": "string"
                },
                "job_skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_job.JobSkill"
                    }
                },
                "job_title": {
                    "type": "string"
                },
                "languages_required": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "no_of_rounds": {
                    "type": "integer"
                },
                "overtime_applicable": {
                    "type": "boolean"
                },
                "pay_schedule": {
                    "type": "string"
                },
                "preference": {
                    "type": "string"
                },
                "remote_location_restriction": {
                    "type": "string"
                },
                "shift_timings": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visa_required": {
                    "type": "boolean"
                },
                "work_mode": {
                    "type": "string"
                },
                "work_type": {
                    "type": "string"
                }
            }
        },
        "internal_job.JobSkill": {
            "description": "Job skill information",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "integer"
                },
                "skill": {
                    "type": "string"
                },
                "type": {
                    "description": "Required, Good-to-have",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}